move: RoomTabs to pages/
